New functionality: Moon Overlay

+++ indi_allsky/timelapse_preprocessor/preProcessorWrapKeogram.py
@@ -128,7 +128,7 @@ class PreProcessorWrapKeogram(PreProcessorBase):
 
         # separate layers
         wrapped_keogram_bgr = wrapped_keogram[:, :, :3]
-        wrapped_keogram_alpha = (wrapped_keogram[:, :, 3] / 255).astype(numpy.float32)
+        wrapped_keogram_alpha = (wrapped_keogram[:, :, 3] / 255).astype(numpy.float16)
 
         # create alpha mask
         alpha_mask = numpy.dstack((
